Fix compilation errors

+++ 0x0B-malloc_free/0-create_array.c
@@ -11,9 +11,9 @@
 char *create_array(unsigned int size, char c)
 {
 	char *str;
-	int i = 0;
+	unsigned int i = 0;
 
-	if (size == 0)
+	if ((int)size == 0)
 	{
 		return (NULL);
 	}
